+++ ruoyi-admin/src/main/java/org/dromara/myCustomer/domain/DcCustomerTransfer.java
@@ -4,6 +4,8 @@ import org.dromara.common.tenant.core.TenantEntity;
 import com.baomidou.mybatisplus.annotation.*;
 import lombok.Data;
 import lombok.EqualsAndHashCode;
+
+import java.math.BigDecimal;
 import java.util.Date;
 import com.fasterxml.jackson.annotation.JsonFormat;
 
@@ -102,12 +104,12 @@ public class DcCustomerTransfer extends TenantEntity {
     /**
      * 实付金额
      */
-    private Long actualPayment;
+    private BigDecimal actualPayment;
 
     /**
      * 尾款情况
      */
-    private Long balanceStatus;
+    private BigDecimal balanceStatus;
 
     /**
      * 签约类型(1-常法 2-单项 3-律师费 4-其他)
